creating a base class for smo scripting operations
